[#25] cleanup: 코드 위치 혹은 스타일 만 변경 로직 변경 없었음
