+++ client/src/utils/utils.js
@@ -1,22 +1,19 @@
 import {getItem} from "./localStorage.js"
+
 export const pathToRegex = path => new RegExp("^" + path.replace(/\//g, "\\/").replace(/:\w+/g, "(.+)") + "$");
 
 export const getParams  = match => {
-    const values = match.result.slice(1);
-    const keys = Array.from(match.route.path.matchAll(/:(\w+)/g)).map(result => result[1]);
+  const values = match.result.slice(1);
+  const keys = Array.from(match.route.path.matchAll(/:(\w+)/g)).map(result => result[1]);
+  return Object.fromEntries(keys.map((key, i) => {
+    return [key, values[i]];
+  }));
+};
 
-    return Object.fromEntries(keys.map((key, i) => {
-      return [key, values[i]];
-    }));
-  };
 export const getParameterByName = name => {
-
     name = name.replace(/[\[]/, "\\[").replace(/[\]]/, "\\]");
-
     const regex = new RegExp("[\\?&]" + name + "=([^&#]*)"),
-
     results = regex.exec(location.search);
-
     return results == null ? "" : decodeURIComponent(results[1].replace(/\+/g, " "));
 
 }
@@ -30,4 +27,3 @@ export const getTagName = tagId => {
 export const formatDate = date => {
   return  new Date(+new Date(date) + 3240 * 10000).toISOString().split("T")[0];
 }
-
